Use flock instead of checking for running pulls using pgrep
The previous solution is potentially racey, it only prevents starting a
pull if there is already one running, but there is still a tiny window
of time where both pulls could start after the checks and run at the
same time. This new solution should fill that gap, since the locking
should be atomic.